Use hardcoded tilestats
